+++ scripts/manager-utilities.js
@@ -87,7 +87,7 @@ export class UtilsManager {
     }
 
     // Sound Management
-    static async playSound(sound = 'sound', volume = 0.7, loop = false, broadcast = true, duration = 0) {
+    static async playSound(sound = 'sound', volume = 0.5, loop = false, broadcast = true, duration = 0) {
         this.checkInitialized();
         return GlobalUtils.playSound(sound, volume, loop, broadcast, duration);
     }